requirements.txt
streamlit
openai>=1.0.0
youtube_transcript_api
langchain
python-dotenv
yt-dlp
requests>=2.31.0
